Exchange: Register always the client cert if present
BZ: 155079

EasServerConnection#executeHttpUriRequest can create a new different connection that the
caller connection, because the use of the new cache connection classes. This occurs after
the client registry was registered in the all connection. To avoid that, this change
moves the registration of the client cert to the cache.

Category: aosp improvement
Domain: aosp-Email
Severity: Medium
Origin: internal
Upstream-Candidate: yes

Orig-Change-Id: If62f99ea81f5951a1d2ee9b692b849cf533507ab
Change-Id: I6035b6f388f7908ce989a898ddf390333e37368c
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>
Reviewed-on: http://android.intel.com:8080/147898
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Wu, GuojingX <guojingx.wu@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>
Tested-by: Balestriere, VianneyX <vianneyx.balestriere@intel.com>